[dat] Introduced the dat::DatasetCompressed type.
It is a major decision shoud we use dynamic strict typing like python, shark, dlib.
Or static strict typing like dlib, this project.

The problem with static typing is changing the shape of the dataset requires a ton of boilerplate code.
The upside is that everything is very clear and readable.

Considering those and that the target audience of the project are extremely junior data scientists (me),
static types were chosen.
They ensure horrible mess when reducing dimensions but are somewhat easier to read to novices.